% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions_ui.R
\name{add_changedate_button}
\alias{add_changedate_button}
\title{Generate +/- button to adjust dates}
\usage{
add_changedate_button(id, legend, class)
}
\arguments{
\item{id}{a \code{character}, used to generate the \code{actionButton} 
\code{id}.}

\item{legend}{a \code{character}, separate the two  \code{actionButton}}

\item{class}{a \code{character}, define the \code{HTML class} used by the
\code{fluidRow}. This class is then used to format the \code{actionButton}}
}
\value{
\code{HTML} code withe two \code{actionButton}
}
\description{
This function generates two \code{actionButton} organized 
within a \code{HTML} \code{fluidRow}. The buttons can then be used within 
the app to adjust the date window.
}
\examples{
add_changedate_button(id = "datemin", legend = "Month", class = "dateclass")


}
\seealso{
Other shiny: 
\code{\link{get_dynamic_input}()},
\code{\link{get_main_title}()},
\code{\link{get_range_input}()},
\code{\link{get_shiny_filename}()},
\code{\link{musica_server}()},
\code{\link{musica_ui}()},
\code{\link{shinymusica}()}
}
\author{
Remi Lemaire-Patin
}
\concept{shiny}
