% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{attr_legend}
\alias{attr_legend}
\title{Extract legend from \code{data.frame} attributes}
\usage{
attr_legend(df)
}
\arguments{
\item{df}{a \code{data.frame}, output of \code{\link{get_variable}}}
}
\value{
A \code{POSIXct} vector
}
\description{
Extract a proper formatted legend from the attributes stored
in a \code{data.frame}
}
\examples{
library(ncdf4)

}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{filter_dim}()},
\code{\link{format_time}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{lat_heat_vap}()},
\code{\link{mol2kg_water_ratio}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
