% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_tools.R
\name{convert.units}
\alias{convert.units}
\title{Standard unit conversion}
\usage{
convert.units(values, from, to, dt = NULL)
}
\arguments{
\item{values}{a \code{numeric}, values to be converted}

\item{from}{a \code{character}, origin units}

\item{to}{a \code{character}, aimed units}

\item{dt}{(\emph{optional}) a \code{numeric}, duration of a time step (in s)
required for some unit conversion}
}
\value{
A \code{numeric}
}
\description{
This function converts values between specific set of units 
(see Details)
}
\details{
The corresponding conversion are available
\itemize{
  \item \code{from = "mmol/m2/dt" ; to = "kg/m2/s"}. Require argument \code{dt}
}
}
\examples{
convert.units(900, from = "mmol/m2/dt", to = "kg/m2/s", dt = 1800)

}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{filter_dim}()},
\code{\link{format_time}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{lat_heat_vap}()},
\code{\link{list_dim_allvar}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()},
\code{\link{var_with_same_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
