% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dygraph_comparison}
\alias{dygraph_comparison}
\title{produce \code{dygraph} for a single variable}
\usage{
dygraph_comparison(
  df,
  pixheight = NULL,
  pixwidth = NULL,
  axisLabelWidth = 75,
  group = "Overview",
  time_range,
  main.title = NULL,
  diffmodels = FALSE
)
}
\arguments{
\item{df}{a \code{data.frame} object, output of \code{\link{get_variable}},
\code{\link{get_variable_comparison}} or  \code{\link{get_two_variables}}.}

\item{pixheight}{height of the \code{dygraph} object in pixel}

\item{pixwidth}{width of the \code{dygraph} object in pixel}

\item{axisLabelWidth}{height of the y axis label}

\item{group}{\code{character}, name of the group to synchronize multiple
dygraphs}

\item{time_range}{(\emph{optional}) a \code{POSIXct vector} with two values:
the time range to extract}

\item{main.title}{(\emph{optional}) a \code{character} with plot title prefix}

\item{diffmodels}{(\emph{optional}, default \code{FALSE}), a
\code{boolean}, if \code{TRUE}, indicates that \code{df} contains the
difference between two models. Plot is then adjusted accordingly}
}
\value{
A \code{dygraph} object
}
\description{
This function plot the output of \code{\link{get_variable}}
with an interactive dygraph app
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
x.list <- list("model1" = x, "model2" = x)
df <- get_variable_comparison(x.list, varname = "Qle")
dygraph_comparison(df,  pixwidth = 600)
}
\seealso{
Other plot: 
\code{\link{dygraph_variable}()},
\code{\link{ggplot_list_var}()},
\code{\link{ggplot_variable}()}
}
\author{
Remi Lemaire-Patin
}
\concept{plot}
