% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_changelog.R
\name{format_changelog}
\alias{format_changelog}
\title{Get information on ncdf dimensions}
\usage{
format_changelog(filename)
}
\arguments{
\item{filename}{a \code{character}, changelog file to reformat}
}
\value{
A \code{data.frame}
}
\description{
This function extract dimension id from the \code{ncdf4} object
}
\examples{
library(ncdf4)
}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{convert.units}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{filter_dim}()},
\code{\link{format_time}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{lat_heat_vap}()},
\code{\link{list_dim_allvar}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()},
\code{\link{var_with_same_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
