% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_time}
\alias{format_time}
\title{format netcdf time into POSIXct}
\usage{
format_time(time, date0)
}
\arguments{
\item{time}{a \code{numeric} in hours since Date0 format$}

\item{date0}{a \code{string}, reference date}
}
\value{
A \code{POSIXct} vector
}
\description{
Reformat the netcdf time "hours since 2009-01-01 00:00:00+00:00"
into a proper POSIXct class
}
\examples{
library(ncdf4)

}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{lat_heat_vap}()},
\code{\link{mol2kg_water_ratio}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
