% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_all_var}
\alias{get_all_var}
\title{extract all variable from a \code{ncdf4} object}
\usage{
get_all_var(x)
}
\arguments{
\item{x}{a \code{ncdf4} object}
}
\value{
A \code{data.frame}
}
\description{
This function extract all variable from a \code{ncdf4} object 
and format it as a nice list of \code{data.frame}.
}
\details{
\describe{
Additional Description
}
}
\examples{
library(ncdf4)

}
\seealso{
Other getters: 
\code{\link{get_variable}()}
}
\author{
Remi Lemaire-Patin
}
\concept{getters}
