% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_dim_info}
\alias{get_dim_info}
\title{Get information on ncdf dimensions}
\usage{
get_dim_info(x)
}
\arguments{
\item{x}{a \code{ncdf4} object}
}
\value{
A \code{data.frame}
}
\description{
This function extract dimension id from the \code{ncdf4} object
}
\examples{
library(ncdf4)

}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{format_time}()},
\code{\link{get_dim_value}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
