% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_dim_value}
\alias{get_dim_value}
\title{Get dimension value from \code{ncdf4} object}
\usage{
get_dim_value(x, dimname)
}
\arguments{
\item{x}{a \code{ncdf4} object}
}
\value{
A \code{vector}
}
\description{
This function extract dimension value from the \code{ncdf4} object
}
\examples{
library(ncdf4)

}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{format_time}()},
\code{\link{get_dim_info}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
