% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions_ui.R
\name{get_dynamic_input}
\alias{get_dynamic_input}
\title{Generate variable selectInput}
\usage{
get_dynamic_input(
  x,
  this.var,
  input.name,
  input.var,
  multiple = FALSE,
  hide = FALSE
)
}
\arguments{
\item{x}{a named list of \code{ncdf4} object. Outputs generated by several
MuSICA model run.}

\item{this.var}{a \code{character}, dimension variable for which to generate 
the  \code{selectInput}. Can be either \code{nsoil}, \code{nair}, 
\code{nveg}, \code{nspecies} or \code{nleafage}.}

\item{input.name}{a \code{character}, reference name of the
\code{selectInput}, used internally in the shiny app to refer to the
results of the input form.}

\item{input.var}{a \code{character}, current variable selected in the app 
for which the \code{selectInput} is generated.}

\item{multiple}{(\emph{optional}, default \code{FALSE}), a \code{boolean},
whether multiple choices are allowed}

\item{hide}{(\emph{optional}, default \code{FALSE}), a \code{boolean}, if 
\code{TRUE}, the \code{selectInput} is not displayed.}
}
\value{
A \code{selectInput} object
}
\description{
The function generates custom \code{selectInput}, used to
  subset variables in the shinymusica app. The \code{selectInput} is
  displayed only if the selected variable (\code{input.var}) matches the
  selected dimension (\code{this.var}).
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
x.list <- list("model1" = x, "model2" = x)
# return selectInput for nsoil
get_dynamic_input(x.list, this.var = "nsoil", input.name = "tab2_nsoil", input.var = "w_soil")
# return NULL as nsoil is not a dimension for Qle
get_dynamic_input(x.list, this.var = "nsoil", input.name = "tab2_nsoil", input.var = "Qle")
}
\seealso{
Other shiny: 
\code{\link{add_changedate_button}()},
\code{\link{get_main_title}()},
\code{\link{get_range_input}()},
\code{\link{get_shiny_filename}()},
\code{\link{musica_server}()},
\code{\link{musica_ui}()},
\code{\link{shinymusica}()}
}
\author{
Remi Lemaire-Patin
}
\concept{shiny}
