% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions_ui.R
\name{get_main_title}
\alias{get_main_title}
\title{plot for a single variable}
\usage{
get_main_title(
  x,
  this.var,
  this.soil,
  this.air,
  this.species,
  this.veg,
  this.leaf
)
}
\arguments{
\item{df}{a \code{data.frame} object}
}
\value{
A \code{ggplot} object
}
\description{
This function plot the output of \code{\link{get_variable}}
}
\examples{
library(ncdf4)


}
\seealso{
Other plot: 
\code{\link{dygraph_comparison}()},
\code{\link{dygraph_variable}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dynamic_input}()},
\code{\link{get_shiny_filename}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{ggplot_list_var}()},
\code{\link{ggplot_variable}()},
\code{\link{musica_server}()},
\code{\link{musica_ui}()},
\code{\link{shinymusica}()}
}
\author{
Remi Lemaire-Patin
}
\concept{plot}
