% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions_ui.R
\name{get_range_input}
\alias{get_range_input}
\title{Generate \code{numericInput} to adjust ranges}
\usage{
get_range_input(tab, type)
}
\arguments{
\item{tab}{a \code{numeric}, 1 or 2 to identify which of the app tab is 
concerned here.}

\item{type}{a \code{character}, aesthetics for which \code{numericInput} 
should be displayed. Can be \code{'x'}, \code{'y'} or \code{'fill'}.}
}
\value{
A set of \code{numericInput} object
}
\description{
This function returns two numeric input presented in a 
\code{fluidRow}. It assigns \code{ID} depending on the given \code{tab} and
the aesthetic concerned (\code{type}).
}
\examples{
get_range_input(1, "fill")
get_range_input(2, "x")
get_range_input(2, "y")
}
\seealso{
Other shiny: 
\code{\link{add_changedate_button}()},
\code{\link{get_dynamic_input}()},
\code{\link{get_main_title}()},
\code{\link{get_shiny_filename}()},
\code{\link{musica_server}()},
\code{\link{musica_ui}()},
\code{\link{shinymusica}()}
}
\author{
Remi Lemaire-Patin
}
\concept{shiny}
