% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions_ui.R
\name{get_shiny_filename}
\alias{get_shiny_filename}
\title{Generate filename for export}
\usage{
get_shiny_filename(var, prefix, suffix, file.format)
}
\arguments{
\item{var}{a \code{character}, currently selected variable}

\item{prefix}{a \code{character}, prefix to add to the filename}

\item{suffix}{a \code{character}, suffix to add to the filename}

\item{file.format}{a \code{character}, file format}
}
\value{
A \code{character}
}
\description{
This function generate a filename used as a suggestion to
  export and download the graphics from tab2. The filename is based around
  the current variable (\code{var}), a \code{prefix}, a \code{suffix} and
  the selected \code{file.format}.
}
\examples{
get_shiny_filename(var = "Qle", prefix = "LeBray", suffix = "crack_depth_1", file.format = "png")
get_shiny_filename(var = "Qle", prefix = "", suffix = "", file.format = "png")

}
\seealso{
Other shiny: 
\code{\link{add_changedate_button}()},
\code{\link{get_dynamic_input}()},
\code{\link{get_main_title}()},
\code{\link{get_range_input}()},
\code{\link{musica_server}()},
\code{\link{musica_ui}()},
\code{\link{shinymusica}()}
}
\author{
Remi Lemaire-Patin
}
\concept{shiny}
