% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_soil_level}
\alias{get_soil_level}
\title{get soil layer bottom and top depth}
\usage{
get_soil_level(x, soil.level)
}
\arguments{
\item{x}{a \code{ncdf4} object. Output generated by MuSICA}

\item{soil.level}{a \code{numeric}, soil level to keep.}
}
\value{
A \code{ggplot} object
}
\description{
This function return bottom and top depth of a single soil
  layer
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
get_soil_level(x, soil.level = 1)
get_soil_level(x, soil.level = 10)
get_soil_level(x, soil.level = 60)
}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{convert.units}()},
\code{\link{e_air}()},
\code{\link{e_air_sat}()},
\code{\link{filter_dim}()},
\code{\link{format_time}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{get_time_range}()},
\code{\link{lat_heat_vap}()},
\code{\link{list_allvar}()},
\code{\link{list_dim_allvar}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()},
\code{\link{var_with_same_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
