% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_species_list}
\alias{get_species_list}
\title{Get simulated species list}
\usage{
get_species_list(x)
}
\arguments{
\item{x}{a \code{ncdf4} object, output of MuSICA}
}
\value{
A \code{character vector}
}
\description{
This function retrieve the secies list from a \code{ncdf4}
 object, output of MuSICA
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
x.list <- list("model1" = x, "model2" = x)
}
\seealso{
Other getters: 
\code{\link{get_all_var}()},
\code{\link{get_timeline}()},
\code{\link{get_timeline_df}()},
\code{\link{get_two_variables}()},
\code{\link{get_variable}()},
\code{\link{get_variable_comparison}()},
\code{\link{get_version}()}
}
\author{
Remi Lemaire-Patin
}
\concept{getters}
