% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions_ui.R
\name{get_subset_input}
\alias{get_subset_input}
\title{Generate variable selectInput for subsetting}
\usage{
get_subset_input(x, this.var, input.name, multiple = FALSE)
}
\arguments{
\item{x}{a named list of \code{ncdf4} object. Outputs generated by several
MuSICA model run.}

\item{this.var}{a \code{character}, dimension variable for which to generate 
the  \code{selectInput}. Can be either \code{nsoil}, \code{nair}, 
\code{nveg}, \code{nspecies} or \code{nleafage}.}

\item{input.name}{a \code{character}, reference name of the
\code{selectInput}, used internally in the shiny app to refer to the
results of the input form.}

\item{multiple}{(\emph{optional}, default \code{FALSE}), a \code{boolean},
whether multiple choices are allowed}
}
\value{
A \code{selectInput} object
}
\description{
The function generates custom \code{selectInput}, used to
  subset variables in the shinymusica app.
}
\author{
Remi Lemaire-Patin
}
