% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{get_timeline}
\alias{get_timeline}
\title{Get a timeline from a \code{ncdf4} object}
\usage{
get_timeline(x, varnames, time_range = NULL)
}
\arguments{
\item{x}{a \code{ncdf4} object. Input data}

\item{varnames}{a \code{character}, the variable to extract}

\item{time_range}{(\emph{optional}) a \code{POSIXct vector} with two values:
the time range to extract}
}
\value{
A \code{data.frame}
}
\description{
This function produce a timeline for available data based on a
  \code{ncdf4} object and a list of variables
}
\examples{
library(ncdf4)
}
\seealso{
Other getters: 
\code{\link{get_all_var}()},
\code{\link{get_two_variables}()},
\code{\link{get_variable_comparison}()},
\code{\link{get_variable}()}
}
\author{
Remi Lemaire-Patin
}
\concept{getters}
