% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{get_timeline_df}
\alias{get_timeline_df}
\title{get a timeline}
\usage{
get_timeline_df(df)
}
\arguments{
\item{df}{a \code{data.frame} object,  with columns \code{time}
(\code{POSIXct}), \code{variable} (\code{character)}) and
\code{has.values} (\code{1} or \code{0})}
}
\value{
A \code{data.frame} object
}
\description{
This function produce a timeline data.frame to feed 
\code{\link{plot_timeline}}
}
\examples{
library(ncdf4)
}
\seealso{
Other getters: 
\code{\link{get_all_var}()},
\code{\link{get_species_list}()},
\code{\link{get_timeline}()},
\code{\link{get_two_variables}()},
\code{\link{get_variable}()},
\code{\link{get_variable_comparison}()},
\code{\link{get_version}()}
}
\author{
Remi Lemaire-Patin
}
\concept{getters}
