% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_variable}
\alias{get_variable}
\title{Get a variable from a \code{ncdf4} object}
\usage{
get_variable(x, varname, time_range, return.colnames = FALSE)
}
\arguments{
\item{x}{a \code{ncdf4} object. Output generated by MuSICA}

\item{varname}{a \code{character}, the variable to extract}

\item{time_range}{(\emph{optional}) a \code{POSIXct vector} with two values:
the time range to extract}

\item{return.colnames}{(\emph{optional}) a \code{boolean}. If \code{TRUE},
return only the column names. If \code{FALSE} return the full
\code{data.frame}}
}
\value{
A \code{data.frame} or a character \code{vector}
}
\description{
This function extract a variable from a \code{ncdf4} object and
format it in a nice \code{data.frame}.
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
get_variable(x, "Qle")
get_variable(x, "root_uptake", return.colnames = TRUE)
time_range <- get_dim_value(x, "time")[c(1, 335)]
get_variable(x, "Qle", time_range = time_range)
}
\seealso{
Other getters: 
\code{\link{get_all_var}()},
\code{\link{get_species_list}()},
\code{\link{get_timeline}()},
\code{\link{get_timeline_df}()},
\code{\link{get_two_variables}()},
\code{\link{get_variable_comparison}()},
\code{\link{get_version}()}
}
\author{
Remi Lemaire-Patin
}
\concept{getters}
