% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_version}
\alias{get_version}
\title{Get MuSICA version number}
\usage{
get_version(x)
}
\arguments{
\item{x}{a \code{ncdf4} object or a list \code{ncdf4} object. (Named list of) 
MuSICA outputs}
}
\value{
A \code{character vector}
}
\description{
This function retrieve MuSICA version number from a single
\code{ncdf4} object or a list of \code{ncdf4} object.
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
x.list <- list("model1" = x, "model2" = x)
}
\seealso{
Other getters: 
\code{\link{get_all_var}()},
\code{\link{get_species_list}()},
\code{\link{get_timeline}()},
\code{\link{get_timeline_df}()},
\code{\link{get_two_variables}()},
\code{\link{get_variable}()},
\code{\link{get_variable_comparison}()}
}
\author{
Remi Lemaire-Patin
}
\concept{getters}
