% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_list_var}
\alias{ggplot_list_var}
\title{plot for a list of variable}
\usage{
ggplot_list_var(x, list_var, time_range)
}
\arguments{
\item{x}{a \code{ncdf4} object. Output generated by MuSICA}

\item{list_var}{a \code{character vector} of variable names to be plotted}

\item{time_range}{(\emph{optional}) a \code{POSIXct vector} with two values:
the time range to extract}
}
\value{
A \code{grid} object with all plots aggregated.
}
\description{
\code{ggplot_list_var} is a wrapper around \code{\link{ggplot_variable}} to produce plots for a list of variables. Plots are aggregated using \code{plot_grid}
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
ggplot_list_var(x, list_var = c("Qle","Qh","Qg"))

}
\seealso{
Other plot: 
\code{\link{dygraph_comparison}()},
\code{\link{dygraph_variable}()},
\code{\link{ggplot_variable}()},
\code{\link{plot_timeline}()}
}
\author{
Remi Lemaire-Patin
}
\concept{plot}
