% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_tools.R
\name{lat_heat_vap}
\alias{lat_heat_vap}
\title{latent heat of vaporisation}
\usage{
lat_heat_vap(tk_air)
}
\arguments{
\item{tk_air}{a \code{numeric}, temperature in Kelvin}
}
\value{
A \code{numeric}
}
\description{
This function computes latent heat of vaporisation (J kg-1) 
from air temperature (K). From Monteith J.L., 1973, Principles of Env. 
Physics, Arnold.
}
\examples{
lat_heat_vap(300)

}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{filter_dim}()},
\code{\link{format_time}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{list_dim_allvar}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()},
\code{\link{var_with_same_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
