% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{list_dim_allvar}
\alias{list_dim_allvar}
\title{List all variables dimensions}
\usage{
list_dim_allvar(x)
}
\arguments{
\item{x}{a \code{ncdf4} object. Output generated by MuSICA}
}
\value{
A \code{list} of \code{vector}
}
\description{
List dimensions for all variables contained in a MuSICA
  \code{netcdf} output.
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
list_dim_allvar(x)
}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{convert.units}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{filter_dim}()},
\code{\link{format_changelog}()},
\code{\link{format_time}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{lat_heat_vap}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()},
\code{\link{var_with_same_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
