\name{musica.tools-package}
\alias{musica.tools-package}
\alias{musica.tools}
\docType{package}
\title{
\packageTitle{musica.tools}
}
\description{
\packageDescription{musica.tools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{musica.tools}
\packageIndices{musica.tools}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{musica.tools}

Maintainer: \packageMaintainer{musica.tools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
