% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.tools-package.R
\docType{package}
\name{musica.tools}
\alias{musica.tools}
\title{musica.tools: tools to extract and plot MuSICA outputs}
\description{
Provides several function to extract MuSICA output from netcdf files.
Output can then be used to produce a series of generic plots.
}
