% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{musica_ui}
\alias{musica_ui}
\title{Shiny UI for \code{\link{shinymusica}}}
\usage{
musica_ui(x)
}
\arguments{
\item{x}{a named list of \code{ncdf4} object. Outputs generated by several
MuSICA model run.}
}
\value{
A \code{shiny ui} object
}
\description{
This function generates a shiny UI for \code{\link{shinymusica}}
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
x.list <- list("model1" = x, "model2" = x)
x.ui <- musica_ui(x.list)

}
\seealso{
Other shiny: 
\code{\link{add_changedate_button}()},
\code{\link{get_dynamic_input}()},
\code{\link{get_main_title}()},
\code{\link{get_range_input}()},
\code{\link{get_shiny_filename}()},
\code{\link{musica_server}()},
\code{\link{shinymusica}()}
}
\author{
Remi Lemaire-Patin
}
\concept{shiny}
