% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{plot a timeline}
\usage{
plot_timeline(df, linewidth = 2, nbreaks = 25)
}
\arguments{
\item{df}{a \code{data.frame} object,  output of \code{\link{get_timeline}}}
}
\value{
A \code{ggplot} object
}
\description{
This function plot a timeline based on the output of
  \code{\link{get_timeline}}
}
\examples{
library(ncdf4)
}
\seealso{
Other plot: 
\code{\link{dygraph_comparison}()},
\code{\link{dygraph_variable}()},
\code{\link{ggplot_list_var}()},
\code{\link{ggplot_variable}()}
}
\author{
Remi Lemaire-Patin
}
\concept{plot}
