% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymusica.R
\name{shinymusica}
\alias{shinymusica}
\title{ShinyApp to explore MuSICA model output}
\usage{
shinymusica(x)
}
\arguments{
\item{x}{a named list of \code{ncdf4} object. Outputs generated by several
MuSICA model run.}
}
\value{
A \code{shiny} application
}
\description{
shinymusica runs a Shiny application to explore multiple MuSICA
model output
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
x.list <- list("model1" = x, "model2" = x)
# Not Run
# shinymusica(x.list)

}
\seealso{
Other shiny: 
\code{\link{add_changedate_button}()},
\code{\link{get_dynamic_input}()},
\code{\link{get_main_title}()},
\code{\link{get_range_input}()},
\code{\link{get_shiny_filename}()},
\code{\link{musica_server}()},
\code{\link{musica_ui}()}
}
\author{
Remi Lemaire-Patin
}
\concept{shiny}
