% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_changelog}
\alias{summary_changelog}
\title{render a html changelog}
\usage{
summary_changelog(
  input.file,
  out.file = "./MuSICA_changelog.html",
  out.dir = "./"
)
}
\arguments{
\item{input.file}{a \code{character}, path to changelog}

\item{out.file}{a \code{character}, path to output file}

\item{out.dir}{(default \code{'./'}), a \code{character}, output directory}
}
\description{
This function render a markdown report for the changelog
}
\examples{
# library(ncdf4)

}
\seealso{
Other summary: 
\code{\link{summary_markdown}()},
\code{\link{summary_pdf}()}
}
\author{
Remi Lemaire-Patin
}
\concept{summary}
