% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_markdown}
\alias{summary_markdown}
\title{render a markdown report}
\usage{
summary_markdown(
  x,
  filename,
  template,
  list_var,
  add_study_site = FALSE,
  time_range,
  n.soil.level = 5,
  n.air.level = 5,
  out.dir = "./",
  ...
)
}
\arguments{
\item{x}{a \code{ncdf4} object}

\item{filename}{a \code{character}, file name to save the report, without
extension (automatically added)}

\item{template}{\emph{(optional)}, a \code{character}, with preset of
variables that can be exported, see Details.}

\item{list_var}{\emph{(optional, ignored if \code{template} argument was
given)} a \code{character vector} with list of variables to be exported.}

\item{add_study_site}{(\emph{optional}, default \code{FALSE}) a
\code{boolean}, if \code{TRUE} adds a leaflet with location of study site}

\item{time_range}{(\emph{optional}) a \code{POSIXct vector} with two values:
the time range to extract}

\item{n.soil.level}{(default \code{5}), a \code{numeric}, number of soil
level to be kept}

\item{n.air.level}{(default \code{5}), a \code{numeric}, number of air level
to be kept}

\item{out.dir}{(default \code{'./'}), a \code{character}, output directory
to store results}

\item{...}{additional parameters transmitted to \code{\link{dygraph_variable}}}
}
\description{
This function render a markdown report for a selection of 
variables. Report includes interactive time-series plot using \code{dygraphs}
package for the selection of variables.
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
summary_markdown(x, "test_summary", template = "Overview")
file.remove("test_summary.html")
}
\seealso{
Other summary: 
\code{\link{summary_pdf}()}
}
\author{
Remi Lemaire-Patin
}
\concept{summary}
