% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_pdf}
\alias{summary_pdf}
\title{render a pdf report}
\usage{
summary_pdf(
  x,
  filename,
  template,
  list_var,
  add_study_site = FALSE,
  time_range,
  file.width,
  file.height,
  daily_heatmap = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{ncdf4} object}

\item{filename}{a \code{character}, file name to save the report, without
extension (automatically added)}

\item{template}{\emph{(optional)}, a \code{character}, with preset of
variables that can be exported, see Details.}

\item{list_var}{\emph{(optional, ignored if \code{template} argument was
given)} a \code{character vector} with list of variables to be exported.}

\item{...}{additional parameters transmitted to \code{\link{ggplot_var}}}
}
\description{
This function render a markdown report for a selection of 
variables. Report includes static plot for a given selection of variable
}
\examples{
library(ncdf4)

}
\seealso{
Other summary: 
\code{\link{summary_markdown}()}
}
\author{
Remi Lemaire-Patin
}
\concept{summary}
