% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions_ui.R
\name{toggle_subset_input}
\alias{toggle_subset_input}
\title{hide or show a subset input}
\usage{
toggle_subset_input(session, input, index.var, tab, x, hide = FALSE)
}
\arguments{
\item{session, }{current shiny session}

\item{input}{a named \code{list} with UI input}

\item{index.var}{a \code{numeric}, from 1 to 3 to identify which of the 
app variable is concerned here}

\item{tab}{a \code{numeric}, 1 or 2 to identify which of the app tab is 
concerned here}

\item{x}{a named list of \code{ncdf4} object. Outputs generated by several
MuSICA model run.}

\item{hide}{(\emph{optional}, default \code{FALSE}), a \code{boolean}, if 
\code{TRUE}, hides all \code{selectInput}}
}
\description{
This function toggle the state of some of the
 \code{selectInput} depending on the server state. It hides the one not 
 needed (which variable is not along the current variable dimension) and 
 show the one needed.
}
\examples{
library(ncdf4)


}
\author{
Remi Lemaire-Patin
}
\concept{shiny
#  x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
#  x.list <- list("model1" = x, "model2" = x)
#  toggle_subset_input(session, input, index.var = 1, tab = 2, x = x.list, hide = FALSE)}
