% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{var_with_dim}
\alias{var_with_dim}
\title{Return variable names with a given dimension}
\usage{
var_with_dim(x, this.dim)
}
\arguments{
\item{x}{a \code{ncdf4} object. Output generated by MuSICA}

\item{this.dim}{a \code{character}, the dimension of interest}
}
\value{
A \code{vector}
}
\description{
Extract from a MuSICA \code{netcdf} output the list of
  variables that share a given dimension.
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
var_with_dim(x, "time")
var_with_dim(x, "nsoil")
}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{convert.units}()},
\code{\link{e_air}()},
\code{\link{e_air_sat}()},
\code{\link{filter_dim}()},
\code{\link{format_time}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{lat_heat_vap}()},
\code{\link{list_allvar}()},
\code{\link{list_dim_allvar}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_same_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
