% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{var_with_same_dim}
\alias{var_with_same_dim}
\title{List variables with same dimensions}
\usage{
var_with_same_dim(x, this.var)
}
\arguments{
\item{x}{a \code{ncdf4} object. Output generated by MuSICA}

\item{this.var}{a \code{character}, the variables for which other variables
with the same dimensions will be returned}
}
\value{
A character \code{vector}
}
\description{
List all variables with the exact same dimensions as a given
  variable.
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
var_with_same_dim(x, this.var = "Qle")
var_with_same_dim(x, this.var = "transpir")
}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{convert.units}()},
\code{\link{e_air}()},
\code{\link{e_air_sat}()},
\code{\link{filter_dim}()},
\code{\link{format_time}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{lat_heat_vap}()},
\code{\link{list_dim_allvar}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
