% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{attr_legend}
\alias{attr_legend}
\title{Extract legend from \code{data.frame} attributes}
\usage{
attr_legend(df)
}
\arguments{
\item{df}{a \code{data.frame}, output of \code{\link{get_variable}}}
}
\value{
A \code{POSIXct} vector
}
\description{
Extract a proper formatted legend from the attributes stored
in a \code{data.frame}
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
df <- get_variable(x, "Qle")
attr_legend(df)
}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{convert.units}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{filter_dim}()},
\code{\link{format_time}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{lat_heat_vap}()},
\code{\link{list_dim_allvar}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()},
\code{\link{var_with_same_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
