% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.fun_testIfIn}
\alias{.fun_testIfIn}
\title{Check that value is within a range of option}
\usage{
.fun_testIfIn(x, values)
}
\arguments{
\item{x}{a scalar value, \code{character} or \code{numeric}}
}
\value{
A \code{boolean}
}
\description{
Generic function to check wether a value is within a range of
option or not. Output appropriate error message if the check failed. Original
code from \code{biomod2} package
}
\examples{
library(ncdf4)

}
\seealso{
Other Tools: 
\code{\link{attr_legend}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{filter_dim}()},
\code{\link{format_time}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{lat_heat_vap}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()}
}
\concept{Tools}
