% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dygraph_comparison}
\alias{dygraph_comparison}
\title{produce \code{dygraph} for a single variable}
\usage{
dygraph_comparison(
  df,
  pixheight = 150,
  pixwidth = 1500,
  axisLabelWidth = 75,
  group = "Overview",
  time_range
)
}
\arguments{
\item{df}{a \code{data.frame} object}

\item{pixheight}{height of the \code{dygraph} object in pixel}

\item{pixwidth}{width of the \code{dygraph} object in pixel}

\item{axisLabelWidth}{height of the y axis label}
}
\value{
A \code{dygraph} object
}
\description{
This function plot the output of \code{\link{get_variable}}
with an interactive dygraph app
}
\examples{
library(ncdf4)


}
\seealso{
Other plot: 
\code{\link{dygraph_variable}()},
\code{\link{ggplot_list_var}()},
\code{\link{ggplot_variable}()}
}
\author{
Remi Lemaire-Patin
}
\concept{plot}
