% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_time}
\alias{format_time}
\title{format netcdf time into POSIXct}
\usage{
format_time(time, date0)
}
\arguments{
\item{time}{a \code{numeric} in hours since Date0 format$}

\item{date0}{a \code{string}, reference date}
}
\value{
A \code{POSIXct} vector
}
\description{
Reformat the netcdf time "hours since 2009-01-01 00:00:00+00:00"
into a proper POSIXct class
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
format_time(x$dim[["time"]]$vals, date0 = x$dim[["time"]]$units)
}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{convert.units}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{filter_dim}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{lat_heat_vap}()},
\code{\link{list_dim_allvar}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()},
\code{\link{var_with_same_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
