% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_6month}
\alias{get_6month}
\title{Get 6 month time range}
\usage{
get_6month(musica.list)
}
\arguments{
\item{musica.list}{a \code{ncdf4} or  a named list \code{ncdf4} object.}
}
\value{
A \code{POSIXct vector}
}
\description{
This function extract the first 6 month time range of a 
\code{ncdf4} or list of \code{ncdf4} objects
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
x.list <- list("model1" = x, "model2" = x)
get_6month(x)
get_6month(x.list)

}
\seealso{
Other Tools: 
\code{\link{.fun_testIfIn}()},
\code{\link{attr_legend}()},
\code{\link{convert.units}()},
\code{\link{e_air_sat}()},
\code{\link{e_air}()},
\code{\link{filter_dim}()},
\code{\link{format_changelog}()},
\code{\link{format_time}()},
\code{\link{get_air_level}()},
\code{\link{get_dim_info}()},
\code{\link{get_dim_value}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{lat_heat_vap}()},
\code{\link{list_dim_allvar}()},
\code{\link{mol2kg_water_ratio}()},
\code{\link{var_with_dim}()},
\code{\link{var_with_same_dim}()}
}
\author{
Remi Lemaire-Patin
}
\concept{Tools}
