% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_all_var}
\alias{get_all_var}
\title{extract all variable from a \code{ncdf4} object}
\usage{
get_all_var(x)
}
\arguments{
\item{x}{a \code{ncdf4} object. Output generated by MuSICA}
}
\value{
A \code{data.frame}
}
\description{
This function extract all variable from a \code{ncdf4} object 
and format it as a nice list of \code{data.frame}.
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
get_all_var(x)
}
\seealso{
Other getters: 
\code{\link{get_two_variables}()},
\code{\link{get_variable_comparison}()},
\code{\link{get_variable}()}
}
\author{
Remi Lemaire-Patin
}
\concept{getters}
