% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions_ui.R
\name{get_main_title}
\alias{get_main_title}
\title{generate proper title for dygraph}
\usage{
get_main_title(
  x,
  this.var,
  this.soil,
  this.air,
  this.species,
  this.veg,
  this.leaf
)
}
\arguments{
\item{x}{a named list of \code{ncdf4} object. Outputs generated by several
MuSICA model run.}

\item{this.var}{a \code{character}, currently selected variable}

\item{this.soil}{a \code{numeric}, currently selected \code{nsoil}}

\item{this.air}{a \code{numeric}, currently selected \code{nair}}

\item{this.species}{a \code{numeric}, currently selected \code{nspecies}}

\item{this.veg}{a \code{numeric}, currently selected \code{nveg}}

\item{this.leaf}{a \code{numeric}, currently selected \code{nleafage}}
}
\value{
A \code{ggplot} object
}
\description{
This function generate a detailed title for the dygraph in 
the first tab of the shiny app. The title include the potential subsetting
of dimension and the depth/height associated to the layer represented in
the plot.
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
x.list <- list("model1" = x, "model2" = x)
get_main_title(x.list, this.var = "w_soil", this.soil = 12,
                       this.air = NULL, this.species = NULL, 
                       this.veg = NULL, this.leaf = NULL)
}
\seealso{
Other shiny: 
\code{\link{add_changedate_button}()},
\code{\link{get_dynamic_input}()},
\code{\link{get_range_input}()},
\code{\link{get_shiny_filename}()},
\code{\link{musica_server}()},
\code{\link{musica_ui}()},
\code{\link{shinymusica}()}
}
\author{
Remi Lemaire-Patin
}
\concept{shiny}
