% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_two_variables}
\alias{get_two_variables}
\title{Get two variable from a \code{ncdf4} object}
\usage{
get_two_variables(x, varnames, time_range, ...)
}
\arguments{
\item{x}{a \code{ncdf4} object. Output generated by MuSICA}

\item{varnames}{a \code{character vector} with two variable names}

\item{time_range}{(\emph{optional}) a \code{POSIXct vector} with two values:
the time range to extract}

\item{...}{additional parameters to be given to
\code{\link{get_variable_comparison}}}
}
\value{
A \code{data.frame}
}
\description{
This function extract a set of two variable from a \code{ncdf4}
  object and format it in a nice \code{data.frame}. Both variables must have
  the same dimensions
}
\examples{
library(ncdf4)
x <- nc_open(system.file("extdata", "musica_out_2006_demo.nc", package = "musica.tools"))
x.list <- list("model1" = x, "model2" = x)
get_two_variables(x, varnames = c("Qle","Qh"))
get_two_variables(x, varnames = c("transpir","gpp"), list.species.level = c("pinus_pinaster"))
get_two_variables(x, varnames = c("h_soil","w_soil"), n.soil.level = 4)
get_two_variables(x.list, varnames = c("h_soil","w_soil"))
}
\seealso{
Other getters: 
\code{\link{get_all_var}()},
\code{\link{get_timeline}()},
\code{\link{get_variable}()},
\code{\link{get_variable_comparison}()}
}
\author{
Remi Lemaire-Patin
}
\concept{getters}
