% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_variable_comparison}
\alias{get_variable_comparison}
\title{Get a variable from a list of \code{ncdf4} object}
\usage{
get_variable_comparison(
  x,
  this_var,
  time_range,
  n.soil.level = NULL,
  list.soil.level = NULL,
  n.air.level = NULL,
  list.air.level = NULL,
  n.species.level = NULL,
  list.species.level = NULL,
  n.veg.level = NULL,
  list.veg.level = NULL,
  n.leafage.level = NULL,
  list.leafage.level = NULL
)
}
\arguments{
\item{x}{a \code{ncdf4} object}
}
\value{
A \code{data.frame}
}
\description{
This function extract a variable from a \code{ncdf4} object and
format it in a nice \code{data.frame}.
}
\details{
\describe{
Additional Description
}
}
\examples{
library(ncdf4)

}
\seealso{
Other getters: 
\code{\link{get_all_var}()},
\code{\link{get_variable}()}
}
\author{
Remi Lemaire-Patin
}
\concept{getters}
