% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_list_var}
\alias{ggplot_list_var}
\title{plot for a single variable}
\usage{
ggplot_list_var(x, list_var, time_range, daily_heatmap = TRUE)
}
\arguments{
\item{df}{a \code{data.frame} object}
}
\value{
A \code{ggplot} object
}
\description{
This function plot the output of \code{\link{get_variable}}
}
\examples{
library(ncdf4)


}
\seealso{
Other plot: 
\code{\link{dygraph_var}},
\code{\link{ggplot_variable}()}
}
\author{
Remi Lemaire-Patin
}
\concept{plot}
