% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_variable}
\alias{ggplot_variable}
\title{plot for a single variable}
\usage{
ggplot_variable(
  df,
  time_range,
  format.date = "\%b \%Y",
  out.type,
  x,
  y,
  color,
  linetype,
  shape,
  fill,
  facet_formula,
  xrange = NULL,
  yrange = NULL,
  fillrange = NULL,
  bin2d = TRUE,
  diffmodels = FALSE,
  nrow.facet = NULL
)
}
\arguments{
\item{df}{a \code{data.frame} object}
}
\value{
A \code{ggplot} object
}
\description{
This function plot the output of \code{\link{get_variable}}
}
\examples{
library(ncdf4)


}
\seealso{
Other plot: 
\code{\link{add_changedate_button}()},
\code{\link{dygraph_comparison}()},
\code{\link{dygraph_variable}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_dynamic_input}()},
\code{\link{get_main_title}()},
\code{\link{get_range_input}()},
\code{\link{get_shiny_filename}()},
\code{\link{get_soil_level}()},
\code{\link{get_subset_input}()},
\code{\link{get_time_range}()},
\code{\link{ggplot_list_var}()},
\code{\link{musica_server}()},
\code{\link{musica_ui}()},
\code{\link{shinymusica}()},
\code{\link{toggle_subset_input}()}
}
\author{
Remi Lemaire-Patin
}
\concept{plot}
