% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_variable}
\alias{ggplot_variable}
\title{plot for a single variable}
\usage{
ggplot_variable(df, heatmap = TRUE, daily_heatmap = TRUE, time_range)
}
\arguments{
\item{df}{a \code{data.frame} object}
}
\value{
A \code{ggplot} object
}
\description{
This function plot the output of \code{\link{get_variable}}
}
\examples{
library(ncdf4)


}
\seealso{
Other plot: 
\code{\link{dygraph_comparison}()},
\code{\link{dygraph_variable}()},
\code{\link{get_6month}()},
\code{\link{get_air_level}()},
\code{\link{get_soil_level}()},
\code{\link{get_time_range}()},
\code{\link{ggplot_list_var}()},
\code{\link{musica_server}()},
\code{\link{musica_ui}()},
\code{\link{shinymusica}()}
}
\author{
Remi Lemaire-Patin
}
\concept{plot}
