% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_markdown}
\alias{summary_markdown}
\title{render a markdown report}
\usage{
summary_markdown(
  x,
  filename,
  template,
  list_var,
  add_study_site = FALSE,
  time_range,
  n.soil.level = 5,
  n.air.level = 5,
  out.dir = "./",
  ...
)
}
\arguments{
\item{x}{a \code{ncdf4} object}

\item{filename}{a \code{character}, file name to save the report, without
extension (automatically added)}

\item{template}{\emph{(optional)}, a \code{character}, with preset of
variables that can be exported, see Details.}

\item{list_var}{\emph{(optional, ignored if \code{template} argument was
given)} a \code{character vector} with list of variables to be exported.}

\item{...}{additional parameters transmitted to \code{\link{dygraph_var}}}
}
\description{
This function render a markdown report for a selection of 
variables. Report includes interactive time-series plot using \code{dygraphs}
package for the selection of variables.
}
\examples{
library(ncdf4)

}
\seealso{
Other summary: 
\code{\link{summary_pdf}()}
}
\author{
Remi Lemaire-Patin
}
\concept{summary}
